Class_03_Run_Level_Global_Level/pyproject.toml
[project]
name = "class-03-run-level-global-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
